RESET signal isn't needed for command control
RESET signal isn't needed for command control. It should be useful when flashing the Flash memory while the WT588D is in operation.

This signal takes 22 ms to complete, which is a long time for a microcontroller. Additionally, this signal generates pops because it causes the audio to drop dramatically.